+++ src/FactChunkMaker.java
@@ -3,7 +3,7 @@
  */
 public class FactChunkMaker extends ChunkMaker {
     @Override
-    public Chunk create(byte[] id, byte[] size) {
+    public Chunk create(int[] id, int[] size) {
         return new FactChunk(id, size);
     }
 }
